Same for cloudNew header stamp.

+++ src/cloud_fusion.cpp
@@ -133,7 +133,7 @@ void streamCallbackRobot1(const sensor_msgs::PointCloud2& cloudRos) {
     pcl::fromPCLPointCloud2(cloudTemp, *cloudNew);
     ROS_INFO("I received a point cloud from Robot 1...");
     ROS_INFO("ROS cloud stamp is: %d", (int) cloudRos.header.stamp.toSec());
-    ROS_INFO("New cloud stamp is: %d", cloudNew->header.stamp);
+    ROS_INFO("New cloud stamp is: %d", (int) cloudNew->header.stamp);
 
     std::string cloudFrame = cloudNew->header.frame_id;
     std::string fixedFrame = "/map";
